# openapi.yaml
openapi: 3.1.0
info:
  title: Family Kanban API
  version: 1.0.0
  description: |
    HTTP contract expected by the Family Kanban frontend (`frontend/src/services`).

    Maps 1:1 to the `KanbanService` interface. Passwordless email + one-time code
    auth (codes emailed via Resend; never returned in API responses). Bearer token
    on every authenticated route.
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
  /auth/code:
    post:
      operationId: requestCode
      tags: [auth]
      summary: Send a one-time sign-in code to an email
      description: |
        Starts passwordless sign-in. Persists a one-time code, then emails it
        via Resend. Unknown emails are fine — the account is created on
        successful verification. The code is never returned in the response.
        No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCodeRequest"
      responses:
        "200":
          description: Code emailed successfully (body is empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCodeResponse"
        "400":
          $ref: "#/components/responses/RuleError"
  /auth/verify:
    post:
      operationId: verifyCode
      tags: [auth]
      summary: Verify the one-time code and establish a session
      description: |
        Validates the code. If the email has no account yet, creates one
        automatically. Returns a session plus an access token for subsequent
        requests. No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          $ref: "#/components/responses/RuleError"
  /auth/session:
    get:
      operationId: currentSession
      tags: [auth]
      summary: Return the current session, or null if not signed in
      description: |
        Optional authentication. With a valid Bearer token, returns the session.
        Without a token (or with an invalid/expired one), returns `null` with
        HTTP 200 — not 401 — so the client can bootstrap cleanly.
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: Current session or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Session"
                  - type: "null"
  /auth/sign-out:
    post:
      operationId: signOut
      tags: [auth]
      summary: End the current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Signed out
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Access token returned by `POST /auth/verify`. Send as
        `Authorization: Bearer <accessToken>` on every authenticated route.
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    RuleError:
      description: Domain / validation rule failure (maps to frontend `RuleError`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable reason suitable for UI toasts
...
    User:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
    Session:
      type: object
      required: [user, activeFamilyId]
      properties:
        user:
          $ref: "#/components/schemas/User"
        activeFamilyId:
          type: ["string", "null"]
    AuthSuccess:
      type: object
      required: [user, activeFamilyId, accessToken]
      description: |
        Session fields plus the Bearer token. Frontend session state uses
        `user` + `activeFamilyId`; the client stores `accessToken` for
        `Authorization` headers.
      properties:
        user:
          $ref: "#/components/schemas/User"
        activeFamilyId:
          type: ["string", "null"]
        accessToken:
          type: string
    RequestCodeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    RequestCodeResponse:
      type: object
      description: |
        Empty object. The one-time code is delivered by email only and is never
        included in the API response.
      additionalProperties: false
    VerifyCodeRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
...